# public/openapi.yaml
openapi: 3.1.0
info:
  title: Cimbra API v1
  version: '2026-09-01'
  description: Versioned financial infrastructure API backed by a persistent double-entry ledger, scoped S2S API keys and signed durable webhooks. This deployment is a sandbox and does not move real funds. Every response includes X-Request-Id and Cimbra-Version; responses after successful API-key authorization also expose X-RateLimit-Limit, X-RateLimit-Remaining and X-RateLimit-Reset. Mutations replayed from an idempotency key include Idempotent-Replayed, while error responses may include Cimbra-Should-Retry so official SDKs can make safe retry decisions.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
...
servers:
  - url: https://cimbra-rose.vercel.app
    description: Persistent sandbox. Does not move real funds.
security: []
paths:
...
  /api/v1/wallets:
    get:
      operationId: listWallets
      summary: List customer wallets with cursor pagination
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/PageLimit' }, { $ref: '#/components/parameters/PageCursor' }]
      responses:
        '200':
          description: Wallet page
          content:
            application/json:
              schema:
                type: object
                required: [data, hasMore, nextCursor]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Wallet' } }
                  hasMore: { type: boolean }
                  nextCursor: { type: [string, 'null'] }
        '403': { description: Missing wallets:read scope or console role }
    post:
      operationId: createWallet
      summary: Open a customer wallet with ledger-backed pockets
      description: Creates one product account per program pocket kind. Does not custody real funds or issue an end-user app.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [programId, customerId, externalReference]
              properties:
                programId: { type: string, format: uuid }
                customerId: { type: string, format: uuid }
                externalReference: { type: string, minLength: 2, maxLength: 100 }
      responses:
        '201': { description: Wallet and pockets created }
        '200': { description: Idempotent replay }
        '400': { description: Invalid wallet payload }
        '403': { description: Missing wallets:write scope or finance.write permission }
        '404': { description: Program or customer not found in tenant }
        '409': { description: Duplicate customer/program, reference or idempotency conflict }
...
  /api/v1/wallets/{id}/transfers:
    post:
      operationId: createWalletPocketTransfer
      summary: Move balance between pockets of the same wallet
      description: Resolves pocket IDs to product accounts and executes a native book transfer with the same risk, holds and transfer.create maker/checker policy.
      security: [{ session: [] }, { bearerApiKey: [] }]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [externalReference, sourcePocketId, destinationPocketId, description, amount, currency]
              properties:
                externalReference: { type: string, minLength: 2, maxLength: 100 }
                sourcePocketId: { type: string, format: uuid }
                destinationPocketId: { type: string, format: uuid }
                description: { type: string, minLength: 2, maxLength: 180 }
                amount:
                  description: Positive decimal major-unit amount. Prefer a string to avoid floating-point serialization.
                  oneOf:
                    - { type: string, pattern: '^\d+(?:\.\d+)?$' }
                    - { type: number, exclusiveMinimum: 0, maximum: 10000000 }
                currency: { type: string, enum: [ARS, USD, MXN, COP, BRL, CLP, PEN] }
                signals: { $ref: '#/components/schemas/RiskSignalsInput' }
      responses:
        '201': { description: Pocket transfer settled or placed in risk review }
        '200': { description: Idempotent replay }
        '202': { description: Maker/checker policy is active and a human approval is pending }
        '400': { description: Invalid payload or same source and destination }
        '403': { description: Missing wallets:write scope or finance.write permission }
        '404': { description: Wallet or pocket not found in tenant }
        '409': { description: Wallet inactive, currency mismatch or idempotency conflict }
        '422': { description: Insufficient available balance or risk decline }
...
components:
  parameters:
    ResourceId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    PageLimit:
      in: query
      name: limit
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    PageCursor:
      in: query
      name: cursor
      required: false
      description: Opaque cursor returned as nextCursor by the previous page.
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Stable unique key for this logical write. Replays return the original resource; reusing it with a different payload returns 409.
      schema: { type: string, minLength: 8, maxLength: 100, pattern: '^[A-Za-z0-9._:-]+$' }
  schemas:
...
    Wallet:
      type: object
      required: [id, programId, programName, programDisplayName, customerId, customerName, externalReference, status, statusReason, pocketCount, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        programId: { type: string, format: uuid }
        programName: { type: string }
        programDisplayName: { type: string }
        customerId: { type: string, format: uuid }
        customerName: { type: string }
        externalReference: { type: string }
        status: { type: string, enum: [active, frozen, closed] }
        statusReason: { type: [string, 'null'] }
        pocketCount: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    RiskSignalsInput:
      type: object
      additionalProperties: false
      properties:
        deviceReference: { type: string, minLength: 2, maxLength: 160, writeOnly: true }
        identityReference: { type: string, minLength: 2, maxLength: 160, writeOnly: true }
        deviceTrust: { type: string, enum: [trusted, unknown, suspicious] }
        identityVerified: { type: boolean }
        ipCountry: { type: string, pattern: '^[A-Za-z]{2}$' }
        countryMismatch: { type: boolean }
...
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-cimbra_session
      description: First-party opaque session token. Only its SHA-256 hash is stored server-side.
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: Cimbra API key
      description: S2S key with a cim_sk_test_ prefix. The plaintext secret is shown once and only its SHA-256 hash is stored.